+++ Profile/func/Load-Module.ps1
@@ -1,4 +1,11 @@
 ## content of .\Load-Module.ps1 ##
+<#PSScriptInfo
+.VERSION 0.3
+.AUTHOR Kai Krutscho
+.PROJECTURI https://www.github.com/Kaimodo/PowerShell-Profile
+.DESCRIPTION
+	Install and Import Module.
+#>
 function Load-Module {
 	<#
 	.SYNOPSIS
@@ -67,6 +74,7 @@ function Load-Module {
 		}
 		catch [Exception] {
             Write-Verbose "$($FunctionName): Process.catch"
+			Write-Host "Error on line $($_.InvocationInfo.ScriptLineNumber): $($_.Exception.Message)"
             Write-Output $_.Exception|format-list -force
         }
 	}